Move get_mod_color to separate util file
