+++ .github/workflows/build.yml
@@ -163,6 +163,7 @@ jobs:
           file-description: "RimSort"
           include-data-files: |
             version.xml=version.xml
+          low-memory: true
         env:
           PYTHONPATH: ${{ steps.pythonpath.outputs.PYTHONPATH }}
